==> 6_lmm_tables.Rmd <==
---
output: pdf_document
classoption: landscape
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
library(kableExtra)
```

```{r data_prep, include=FALSE}
#Linear models to explain C storage ============================================
#Adam Mahood and Dr. R. Chelsea Nagy
#updated September 2020

#load multiple libraries 
x <- c("tidyverse", "sf","scales", "ggplot2", "doBy", "lme4", "lmerTest", "stargazer","ggpubr")
lapply(x, library, character.only = TRUE, verbose = FALSE)
theme_set(theme_classic())

siwf <- as.data.frame(read_csv("siwf.csv"))%>%
  filter(veg != 'salt_desert') %>%
  mutate(firecat = ifelse(timesincefire < 5,
                          "< 5", ifelse(timesincefire >= 5 &
                                             timesincefire < 20, "5-20", 
                                           ifelse(timesincefire >= 20, "> 20", NA))),
         firecat = factor(firecat, levels = c("< 5", "5-20", "> 20")),
         veg = factor(veg, levels = c("sagebrush", "sagecheat", "cheatgrass"),
                      labels = c("Native Sagebrush","Invaded Sagebrush","Cheatgrass")))%>%
  mutate(soil_depth = ifelse(bottomdepth_cm <= 10, "0-10", 
                           ifelse(bottomdepth_cm > 10 & bottomdepth_cm <=20, "10-20", 
                                  ifelse(bottomdepth_cm > 20, "20-160", NA))))

colours <- c("Native Sagebrush" = "seagreen4", "Invaded Sagebrush" = "yellowgreen", "Cheatgrass" = "gold")
#try with subsets for each pool
AGBC3 <- subset.data.frame(siwf, pool == "AGBC_g_m2") %>%
  mutate("log(AGBC+1)" = log(pool_value+1)) 
BGBC3 <- subset.data.frame(siwf, pool == "BGBC_g_m2") %>%
  mutate("log(BGBC+1)" = log(pool_value+1))
litterC3 <- subset.data.frame(siwf, pool == "litterC_g_m2")%>%
  mutate("Litter C" = pool_value)
orgsoilC3 <- subset.data.frame(siwf, pool == "orgsoilC_g_m2")%>%
  mutate("log(Soil Organic C)" = log(pool_value))
totsoilC3 <- subset.data.frame(siwf, pool == "totsoilC_g_m2")%>%
  mutate("Soil Total C" = pool_value) 

final_mods <- list()

final_mods$ABG <- lmer(`log(AGBC+1)` ~ veg * firecat + (1|Article_ID), 
                   data=AGBC3, REML = TRUE)
final_mods$BGB <- lmer(`log(BGBC+1)` ~ timesincefire +(timesincefire|Article_ID), 
                    data=BGBC3, REML = TRUE)
final_mods$litter <- lmer(`Litter C` ~ timesincefire + (1|Article_ID), 
                       data=litterC3, REML = TRUE)
final_mods$org_s <- lmer(`log(Soil Organic C)` ~ firecat * veg* soil_depth + (1|Article_ID),
                        data=orgsoilC3, REML=TRUE)
# final_mods$org_s1 <- lmer(`log(Soil Organic C)` ~ veg*soil_depth + veg*firecat +(1|Article_ID),
#                         data=orgsoilC3, REML=FALSE)
# AIC(final_mods$org_s, final_mods$org_s1)
# summary(final_mods$org_s)
final_mods$tot <- lmer(log(`Soil Total C`) ~ timesincefire * veg* bottomdepth_cm + (1|Article_ID), 
                        data=totsoilC3, REML=TRUE)
# summary(final_mods$tot)
```

```{r}
p4a<-ggplot(AGBC3, aes(y=pool_value, fill=veg, x=firecat)) + 
  geom_boxplot() +
  ylab(expression(Aboveground~Biomass~C~(g~m^-2)))+
  xlab("Time Since Fire (Years)") +
  scale_y_log10(labels = scales::label_number_si())+
  scale_fill_manual(values = colours,
                     name = "Vegetation Type")+
  theme(legend.position = "none")
```

```{r}
p4b<-ggplot(BGBC3, aes(y=pool_value, x=timesincefire)) +  
  geom_smooth(method="lm", color = "black") +
  geom_point(aes(color = veg)) +
  scale_y_log10(labels = scales::label_number_si())+
  scale_color_manual(values = colours,
                     name = "Vegetation Type")+
  ylab(expression(Belowground~Biomass~C~(g~m^-2)))+
  xlab("Time Since Fire (Years)") +
  theme(legend.position = "none")

```

```{r}
p4c<-ggplot(litterC3, aes(y=`Litter C`, x=timesincefire)) + 
  geom_smooth(method="lm", color = "black")+
  geom_point(aes(color = veg)) +
  scale_color_manual(values = colours,
                     name = "Vegetation Type")+
  ylab(expression(Litter~C~(g~m^-2))) +
  xlab("Time Since Fire (Years)") +
  theme(legend.position = "none")
```

```{r fig.width=9}
p4d<-orgsoilC3 %>%
   mutate(soil_depth = paste("Depth:", soil_depth, "cm")) %>%
ggplot(aes(y=pool_value, x =firecat, fill = veg)) + 
  geom_boxplot() +
  scale_fill_manual(values = colours,
                     name = "Vegetation Type")+
  facet_wrap(~soil_depth)+
  scale_y_log10(labels = scales::label_number_si())+
  ylab(expression(Soil~Organic~C~(g~m^-2)))+
  xlab("Time Since Fire (Years)")+
  theme(legend.position = c(0,0),
        legend.title = element_blank(),
        legend.justification = c(0,0),
        legend.background = element_rect(fill=NA, color = "black"))
```

```{r}
p4e<-totsoilC3 %>%
  ggplot(aes(y=pool_value, x = timesincefire, color=veg)) +
  geom_smooth(method="lm") +
  geom_point() +
  scale_y_log10(labels = scales::label_number_si())+
  ylab(expression(Total~Soil~C~(g~m^-2)))+
  xlab("Time Since Fire (Years)") +
  scale_color_manual(values = colours,
                     name = "Vegetation Type") +
  theme(legend.position = "none")
```

```{r results='asis'}
rbind(MuMIn::r.squaredGLMM(final_mods$ABG) %>% as.data.frame,
      MuMIn::r.squaredGLMM(final_mods$BGB) %>% as.data.frame,
      MuMIn::r.squaredGLMM(final_mods$litter)%>% as.data.frame,
      MuMIn::r.squaredGLMM(final_mods$org_s)%>% as.data.frame,
      MuMIn::r.squaredGLMM(final_mods$tot)%>% as.data.frame) %>%
  cbind(.,data.frame(model = c("Aboveground Biomass C ~ Invasion Stage * Time Since Fire + (1 | Article ID)",
                                "Belowground Biomass C ~ Time Since Fire + (Time Since Fire | Article ID)",
                                "Litter Carbon ~ Time Since Fire + (1 | Article ID)", 
                                "Soil Organic C ~ Time Since Fire * Invasion Stage * Soil Depth + (1 | Article ID)",
                                "Soil Total C ~ Time Since Fire * Invasion Stage * Soil Depth +      (1 | Article ID)"))) %>%
  dplyr::select(model, R2m, R2c) %>%
  mutate(R2m = round(R2m,2),
         R2c = round(R2c,2)) %>%
  kable(booktabs=T, 
        caption = "Marginal and Conditional pseudo R2 values for linear mixed models") %>%
  kableExtra::footnote(c("R2m: Marginal R-squared", "R2c: Conditional R-squared"))
  # stargazer(summary=F, header=F, 
  #           title = "Marginal and Conditional pseudo R2 values for linear mixed models",
  #           rownames = FALSE)
```

```{r fig.width = 10, fig.height=7.5}
row1<-ggarrange(p4e, p4b, p4c, nrow=1, labels = "auto", label.x = .2)
row2<-ggarrange(p4d, p4a, nrow=1, widths=c(2,1), labels = c("d","e"),
                label.x = c(.1,.2), label.y= c(.9, 1))
ggarrange(row1, row2, nrow=2) +
  ggsave("figure_3.png", height = 8, width =10) +
  ggsave("figure_3.pdf", height = 8, width =10)
```

```{r results='asis'}
for(i in 1:length(final_mods)) class(final_mods[[i]]) <- "lmerMod"
# stargazer(final_mods[c(1:3)], header=F,column.sep.width = "1pt", single.row = T, 
#           title = "Linear mixed model results.")

lut_vars <-c("veg" = "Vegetation Type",
             "firecat" = "Time Since Fire",
             "veg:firecat" = "Vegetation Type x Time Since Fire",
             "timesincefire" = "Time Since Fire",
             "soil_depth" = "Soil Depth",
             "firecat:veg" = "Time Since Fire x Vegetation Type",
             "firecat:soil_depth" = "Time Since Fire x Soil Depth",
             "veg:soil_depth" = "Vegetation Type x Soil Depth",
             "firecat:veg:soil_depth" =   "Time Since Fire x Soil Depth",
             "bottomdepth_cm" = "Soil Depth",
             "timesincefire:veg"  = "Time Since Fire x Vegetation Type",
             "timesincefire:bottomdepth_cm" = "Time Since Fire x Soil Depth",
             "veg:bottomdepth_cm" = "Vegetation Type x Soil Depth",
             "timesincefire:veg:bottomdepth_cm"="Time Since Fire x Vegetation Type x Soil Depth")

aovs<- rbind(
final_mods[[1]] %>% car::Anova()%>% round(4) %>% as_tibble(rownames = "Variable") %>% mutate(var = "Aboveground Biomass C"),
final_mods[[2]] %>% car::Anova()%>% round(4) %>% as_tibble(rownames = "Variable") %>% mutate(var = "Belowground Biomass C"),
final_mods[[3]] %>% car::Anova()%>% round(4) %>% as_tibble(rownames = "Variable") %>% mutate(var = "Litter C"),
final_mods[[4]] %>% car::Anova()%>% round(4) %>% as_tibble(rownames = "Variable") %>% mutate(var = "Soil Organic C"),
final_mods[[5]] %>% car::Anova()%>% round(4) %>% as_tibble(rownames = "Variable") %>% mutate(var = "Soil Total C")
) %>%
  mutate(Variable=lut_vars[Variable])


aovs %>%
  mutate(`p-value` = ifelse(`Pr(>Chisq)` < 0.05, str_c(`Pr(>Chisq)`, " *"), `Pr(>Chisq)`))%>%
  dplyr::select(-`Pr(>Chisq)`, -var) %>%
  kable(booktabs=T, caption = "Linear mixed model results. Significance calculated using type II Wald Chi-squared tests") %>%
  kableExtra::group_rows(index = table(aovs$var))
  
```

```{r results='asis'}
stargazer(final_mods[c(1)], header=F,column.sep.width = "1pt", single.row = T,
          title = "Table 2a: Linear mixed model results.")#,
          # rownames = c("Invaded Sagebrush",
          #              "Cheatgrass",
          #              "Time Since Fire (5-20)",
          #              "Time Since Fire (>20)",
          #              "Invaded Sagebrush x Time Since Fire (5-20)",
          #              "Cheatgrass x Time Since Fire (5-20)",
          #              "Invaded Sagebrush x Time Since Fire (>20)",
          #              "Cheatgrass x Time Since Fire (>20)",
          #              "Constant"))


```

```{r results='asis'}
stargazer(final_mods[c(2,3)], header=F,column.sep.width = "1pt", single.row = T,
          title = "Table 2a: Linear mixed model results.")#,
          # rownames = c("Invaded Sagebrush",
          #              "Cheatgrass",
          #              "Time Since Fire (5-20)",
          #              "Time Since Fire (>20)",
          #              "Invaded Sagebrush x Time Since Fire (5-20)",
          #              "Cheatgrass x Time Since Fire (5-20)",
          #              "Invaded Sagebrush x Time Since Fire (>20)",
          #              "Cheatgrass x Time Since Fire (>20)",
          #              "Constant"))


```

```{r results='asis'}
stargazer(final_mods[c(4)], header=F,column.sep.width = "1pt", single.row = T,
          title = "Table 2b: Linear mixed model results.")
```

```{r results='asis'}
stargazer(final_mods[c(5)], header=F,column.sep.width = "1pt", single.row = T,
          title = "Table 2c: Linear mixed model results.")
```
